move tests back to `send_cmd.rs`
