--- forest/cli/tests/send_tests.rs
@@ -1,161 +1,0 @@
-// Copyright 2019-2023 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-#[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use forest_cli::cli::{
-        send_cmd::FILAmount,
-        wallet_cmd::{format_balance_string, FormattingMode},
-    };
-    use fvm_shared::econ::TokenAmount;
-    use quickcheck_macros::quickcheck;
-
-    #[test]
-    fn invalid_attofil_amount() {
-        //attoFIL with fractional value fails (fractional FIL values allowed)
-        let amount = "1.234attofil";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-
-    #[test]
-    fn valid_attofil_amount_test1() {
-        //valid attofil amount passes
-        let amount = "1234 attofil";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_atto(1234)
-        );
-    }
-
-    #[test]
-    fn valid_attofil_amount_test2() {
-        //valid attofil amount passes
-        let amount = "1234 afil";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_atto(1234)
-        );
-    }
-
-    #[test]
-    fn valid_attofil_amount_test3() {
-        //valid attofil amount passes
-        let amount = "1234 a fil";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_atto(1234)
-        );
-    }
-
-    #[test]
-    fn valid_attofil_amount_test4() {
-        //valid attofil amount passes
-        let amount = "1234 a";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_atto(1234)
-        );
-    }
-
-    #[test]
-    fn suffix_with_no_amount() {
-        //fails if no amount specified
-        let amount = "fil";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-    #[test]
-    fn valid_fil_amount_without_suffix() {
-        //defaults to FIL if no suffix is provided
-        let amount = "1234";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_whole(1234)
-        );
-    }
-
-    #[test]
-    fn valid_fil_amount_with_suffix() {
-        //properly parses amount with "FIL" suffix
-        let amount = "1234FIL";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_whole(1234)
-        );
-    }
-
-    #[test]
-    fn invalid_fil_amount() {
-        //bad amount fails
-        let amount = "0.0.0FIL";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-
-    #[test]
-    fn test_fractional_fil_amount() {
-        //properly parses fil with fractional value
-        let amount = "1.234FIL";
-        assert_eq!(
-            FILAmount::from_str(amount).unwrap().value,
-            TokenAmount::from_atto(1_234_000_000_000_000_000i64)
-        );
-    }
-
-    #[test]
-    fn fil_amount_too_long() {
-        //fil amount with length>50 fails
-        let amount = "100000000000000000000000000000000000000000000000000FIL";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-
-    #[test]
-    fn convert_fil_to_attofil() {
-        //expected attofil amount matches actual amount after conversion from FIL
-        let fil_amount = "1FIL";
-        assert_eq!(
-            FILAmount::from_str(fil_amount).unwrap().value,
-            TokenAmount::from_whole(1)
-        );
-    }
-
-    #[test]
-    fn invalid_fil_suffix() {
-        //fails with bad suffix
-        let amount = "42fiascos";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-
-    #[test]
-    fn malformatted_fil_suffix_test() {
-        //fails with bad suffix
-        let amount = "42 fem to fil";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-
-    #[test]
-    fn negative_fil_value() {
-        //fails with negative value
-        let amount = "-1FIL";
-        assert!(FILAmount::from_str(amount).is_err());
-    }
-
-    #[quickcheck]
-    fn fil_quickcheck_test(n: u64) {
-        let token_amount = TokenAmount::from_atto(n);
-        let formatted =
-            format_balance_string(token_amount.clone(), FormattingMode::ExactNotFixed).unwrap();
-        let parsed = FILAmount::from_str(&formatted).unwrap().value;
-        assert_eq!(token_amount, parsed);
-    }
-
-    #[quickcheck]
-    fn scaled_fil_quickcheck_test(n: u64, rand_num: u32) {
-        let scaled_n = n % u64::pow(10, rand_num % 20);
-        let token_amount = TokenAmount::from_atto(scaled_n);
-        let formatted =
-            format_balance_string(token_amount.clone(), FormattingMode::ExactNotFixed).unwrap();
-        let parsed = FILAmount::from_str(&formatted).unwrap().value;
-        assert_eq!(token_amount, parsed);
-    }
-}
